Create SECURITY.md
To manage the  Core Team Pi Network
Greetings to all of you for the great Pi Network project that serves humanity everywhere in the world without discrimination and creates a new, innovative encryption world that serves all areas of life for the individual and society.
Since I am an Arab member in this project, I would like to inform you of some of the problems of the Arab community in many issues related to the Pi Network project. The most important of these issues is the issue of the theft of Pi Coins belonging to some Arab pioneers, as they are stolen in various ways, and the most important of these methods is the presence of scam applications that request passwords. 24 for the Pi Browser wallet, as the Arab community of the Pi Network project was initially unaware of such serious issues, but after accepting the documentation and transferring the currencies to the wallet, they are stolen to external wallets, and the proof is on the Blockchain Pi, and they are illegally traded on applications or direct dealings, before the eyes of the right holders. And before your eyes.
We would like you to freeze all accounts that dealt with suspicious wallets in the scam and return the balances of the members who were stolen before opening the main network to the Pi Network application and inform them of this in order to change the passwords for the Pi Browser wallet and obtain your balances again.
All greetings, appreciation and respect from the Arab community to you